requirements.txt
selenium==4.14.0
PyYAML~=6.0.0